+++ gamma_dashboard/static/dashboard/js/app/src/components/dashboard/rowBlock/__tests__/useStatusesBlock.unit.test.js
@@ -1,17 +1,19 @@
 import { renderHook } from '@testing-library/react-hooks';
 import { cleanup } from '@testing-library/react';
-import DataUseStatusesBlock from './__mock__/DataUseStatusesBlock.json';
 import { useStatusesBlock } from '../statusesBlock/useStatusesBlock';
+import DataForTestsCurrentIndex from './__mock__/DataForTestsCurrentIndex.json';
+import DataUseStatusesBlock from './__mock__/DataUseStatusesBlock.json';
 
 afterEach(cleanup);
 
-describe('getProgressTrackStyles returns expected styles', () => {
+describe('getProgressTrackStyles', () => {
 
   beforeEach(() => {
     cleanup();
   });
 
-  it.each(DataUseStatusesBlock)('checking different cases', ({ statusItems, index, expectResult }) => {
+  it.each(DataUseStatusesBlock)(
+    'getProgressTrackStyles returns expected styles', ({ statusItems, index, expectResult }) => {
     const { result } = renderHook(() => useStatusesBlock({ statusItems }));
     const styles = result.current.getProgressTrackStyles(index);
 
@@ -19,5 +21,20 @@ describe('getProgressTrackStyles returns expected styles', () => {
     expect(styles.badgeStyles.opacity).toBe(expectResult.opacity);
     expect(styles.progressTrackStyles.width).toBe(expectResult.width);
     expect(styles.progressTrackEndStyles.display).toBe(expectResult.display);
+  });
+});
+
+describe('calculateCurrentIndex', () => {
+
+  beforeEach(() => {
+    cleanup();
+  });
+
+  it.each(DataForTestsCurrentIndex)(
+    'calculateCurrentIndex returns expected index', ({ statusItems, expectResult }) => {
+    const { result } = renderHook(() => useStatusesBlock({ statusItems }));
+    const receivedIndex = result.current.calculateCurrentIndex(statusItems);
+
+    expect(receivedIndex).toBe(expectResult.currentIndex);
   });
 });